Fix pre-patch attribute import
If a submodule is imported as an attribute (i.e., `from foo import bar`)
prior to it being explicitly patched or auto-loaded, the attribute
import ended up getting a different mock instance than the explicit or
auto-loaded import.

This also removes the MockPackage class because we can't actually know
ahead of time which attributes should be treated like a package or not.