Patch charms.layer.import_layer_libs
This function is normally handled automatically but actions have to
manually call it, so we need to ensure it's available.

Fixing this uncovered the reason why the patched modules were losing
their attribute attachment to their parents: `patch.dict()` was
restoring the hidden patched modules but was losing the attribute
attachment, and this could actually affect things other than the
auto-importer, so now it's fixed more generally.