sort the modules we want to hide (#13)
py35 travis was failing because we might attempt to setattr on something
like 'charmhelpers.foo' when 'charmhelpers' itself isn't in sys.modules
yet. Fix it by sorting the modules-to-be-patched so foo comes before
foo.bar.

Not a problem with newer pythons because of the dict insertion order
preservation, but it doesn't hurt to always sort.